requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.1.0
ffmpeg-python==0.2.0
filelock==3.10.0
frozenlist==1.3.3
future==0.18.3
huggingface-hub==0.13.2
idna==3.4
Jinja2==3.1.2
llvmlite==0.39.1
MarkupSafe==2.1.2
more-itertools==9.1.0
mpmath==1.3.0
multidict==6.0.4
networkx==3.0
numba==0.56.4
numpy==1.23.5
openai==0.27.2
openai-whisper==20230314
packaging==23.0
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
sympy==1.11.1
tiktoken==0.3.1
tokenizers==0.13.2
torch==2.0.0
tqdm==4.65.0
transformers==4.27.1
typing_extensions==4.5.0
urllib3==1.26.15
yarl==1.8.2
